requirements.txt
anthropic==0.72.0
python-dotenv==1.0.0
